refactor: change the location of the GetPokemons component
