+++ src/components/ErrorBoundary.tsx
@@ -25,6 +25,7 @@ export default class ErrorBoundary extends Component<
   }
 
   componentDidCatch(error: Error, errorInfo: ErrorInfo) {
+    // eslint-disable-next-line no-console
     console.log(error, errorInfo);
   }
